+++ lib/database.py
@@ -2,14 +2,14 @@
 # Copyright 2023-2025 David Kneipp <david@davidkneipp.com>
 # Copyright 2024-2025 Lennart Rosam <hello@takuto.de>
 # Copyright 2024-2025 Alexander Couzens <lynxis@fe80.eu>
+# Copyright 2025 sysmocom - s.f.m.c. GmbH <info@sysmocom.de>
 # SPDX-License-Identifier: AGPL-3.0-or-later
 from typing import Optional
 
+import sqlalchemy
 from sqlalchemy import Column, Integer, String, MetaData, Table, Boolean, ForeignKey, select, UniqueConstraint, DateTime, BigInteger, Text, DateTime, Float
-from sqlalchemy import create_engine, inspect
-from sqlalchemy.engine.reflection import Inspector
+from sqlalchemy import create_engine
 from sqlalchemy.sql import desc, func
-from sqlalchemy_utils import database_exists, create_database
 from sqlalchemy.orm import sessionmaker, relationship, Session, class_mapper
 from sqlalchemy.orm.attributes import History, get_history
 from sqlalchemy.orm import declarative_base
@@ -22,6 +22,7 @@ import uuid
 import socket
 import pprint
 import S6a_crypt
+from databaseSchema import DatabaseSchema
 from baseModels import SubscriberInfo, LocationInfo2G
 from gsup.protocol.ipa_peer import IPAPeerRole
 from messaging import RedisMessaging
@@ -32,6 +33,13 @@ from pyhss_config import config
 
 
 Base = declarative_base()
+
+class DATABASE_SCHEMA_VERSION(Base):
+    __tablename__ = "database_schema_version"
+    upgrade_id = Column(Integer, primary_key=True, doc="Schema version")
+    comment = Column(String(512), doc="Notes about this version upgrade")
+    date = Column(DateTime(timezone=True), server_default=sqlalchemy.sql.func.now(), doc="When the upgrade was done")
+
 class APN(Base):
     __tablename__ = 'apn'
     apn_id = Column(Integer, primary_key=True, doc='Unique ID of APN')
@@ -356,7 +364,7 @@ class SUBSCRIBER_ATTRIBUTES_OPERATION_LOG(OPERATION_LOG_BASE):
 
 class Database:
 
-    def __init__(self, logTool, redisMessaging=None):
+    def __init__(self, logTool, redisMessaging=None, main_service: bool = False):
 
         self.redisUseUnixSocket = config.get('redis', {}).get('useUnixSocket', False)
         self.redisUnixSocketPath = config.get('redis', {}).get('unixSocketPath', '/var/run/redis/redis-server.sock')
@@ -395,13 +403,7 @@ class Database:
             pool_size=config['logging'].get('sqlalchemy_pool_size', 30),
             max_overflow=config['logging'].get('sqlalchemy_max_overflow', 0))
 
-        # Create database if it does not exist.
-        if not database_exists(self.engine.url):
-            self.logTool.log(service='Database', level='debug', message="Creating database", redisClient=self.redisMessaging)
-            create_database(self.engine.url)
-            Base.metadata.create_all(self.engine)
-        else:
-            self.logTool.log(service='Database', level='debug', message="Database already created", redisClient=self.redisMessaging)
+        DatabaseSchema(self.logTool, Base, self.engine, main_service)
 
         #Load IMEI TAC database into Redis if enabled
         if self.tacDatabasePath:
@@ -410,15 +412,6 @@ class Database:
         else:
             self.logTool.log(service='Database', level='info', message="Not loading EIR IMEI TAC Database as Redis not enabled or TAC CSV Database not set in config", redisClient=self.redisMessaging)
             self.tacData = {}
-
-        # Create individual tables if they do not exist
-        inspector = inspect(self.engine)
-        for table_name in Base.metadata.tables.keys():
-            if table_name not in inspector.get_table_names():
-                self.logTool.log(service='Database', level='debug', message=f"Creating table {table_name}", redisClient=self.redisMessaging)
-                Base.metadata.tables[table_name].create(bind=self.engine)
-            else:
-                self.logTool.log(service='Database', level='debug', message=f"Table {table_name} already exists", redisClient=self.redisMessaging)
 
     def load_IMEI_database_into_Redis(self):
         try:
@@ -1646,7 +1639,8 @@ class Database:
         elif action == "2g3g":
             # Mask first bit of AMF
             key_data['amf'] = '0' + key_data['amf'][1:]
-            vect = S6a_crypt.generate_2g3g_vector(key_data['ki'], key_data['opc'], key_data['amf'], int(key_data['sqn']), int(key_data['algo']))
+            algo = int(key_data["algo"]) if key_data["algo"] is not None else 3
+            vect = S6a_crypt.generate_2g3g_vector(key_data['ki'], key_data['opc'], key_data['amf'], int(key_data['sqn']), algo)
             vector_list = []
             self.logTool.log(service='Database', level='debug', message="Generating " + str(kwargs['requested_vectors']) + " vectors for GSM use", redisClient=self.redisMessaging)
             while kwargs['requested_vectors'] != 0: